Add slugs back to all examples (#4)
Now that the `init` command creates slugs and deploy expects a `slug` field to be set, we will need to pre-configure slugs for all of our examples. However, that works out well since that means all our examples are now valid rather than relying on our CLI to add a slug for you.

In a future CLI PR though, we'll need to add logic to the `deploy` command to check if the slug is unique and ask if the user wants to create a new task with a different slug.